refactor(client): simplify dns interceptor lazy initialization
